Add Mach Zehnder
new freecad models: 1inch adapter, Manta allied vision camera
new moduls: beamsplitter (from Clemens test branch), export to tikz, polarization rotator

new file: Mach Zehnder Interferometer
