+++ LaserCAD/basic_optics/__init__.py
@@ -29,3 +29,5 @@ from .mount import Unit_Mount, Grating_Mount, Composed_Mount, Stripe_Mirror_Moun
 from .post import Post_and_holder
 from .refractive_plane import Refractive_plane
 from .off_axis_parabola import Off_Axis_Parabola
+from .composition import export_to_TikZ, print_post_positions
+from .beam_splitter import ThinBeamsplitter, ThickBeamsplitter, TFP56
