+++ LaserCAD/freecad_models/mount_meshes/mirror/mirrormounts.csv
@@ -43,3 +43,5 @@ U200-A2K,50.8,25.4,0,-35.052,0,0,1,"0.2,0.2,0.2"
 U300-A2K,76.2,25,0,-48,0,0,1,"0.2,0.2,0.2"
 ,,,,,,,,
 Optosigma_MHGT-50_8,50.8,20,0,-35.5,0,0,1,"0.2,0.2,0.2"
+,,,,,,,,
+Newport-9771-M,25.4,18.7,0,-16,0,0,1,"0.2,0.2,0.2"
