+++ LaserCAD/non_interactings/__init__.py
@@ -13,4 +13,5 @@ from .crystal import Crystal
 from .breadboard import Breadboard
 from .table import Table
 from .laser_pointer import LaserPointer
+from .camera import Camera
 # from .mount import Unit_Mount, Grating_Mount, Composed_Mount, Stripe_Mirror_Mount, Rooftop_Mirror_Mount, Post
