+++ LaserCAD/non_interactings/pockels_cell.py
@@ -10,9 +10,9 @@ from ..basic_optics.component import Component
 from ..freecad_models.utils import thisfolder, load_STL
 
 class Pockels_Cell(Component):
-  def __init__(self, name="Pockels_Cell", **kwargs):
+  def __init__(self, name="Pockels_Cell", mount_name = "pockels_cell_easy_steal-Body", **kwargs):
     super().__init__(name, **kwargs)
-    stl_file=thisfolder+"misc_meshes/pockels_cell_easy_steal-Body.stl"
+    stl_file=thisfolder+"misc_meshes/"+mount_name+".stl"
     self.draw_dict["stl_file"]=stl_file
     self.draw_dict["color"]=(239/255, 239/255, 239/255)
     # self.draw_dict["color"]=(100/255, 100/255, 100/255)
