--- moduls/__init__.py
@@ -1,15 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Thu Jun 22 10:58:57 2023
-
-@author: 12816
-"""
-
-from .telescope import Make_Telescope
-from .periscope import Make_Periscope, Make_RoofTop_Mirror
-from .stretcher import Make_Stretcher, Make_Stretcher_chromeo
-from .type_I_Amplifier import Make_Amplifier_Typ_I_Mirror, Make_Amplifier_Typ_I_simple
-from .type_II_Amplifier import Make_Amplifier_Typ_II_simple, Make_Amplifier_Typ_II_Mirror
-from .type_II_Amplifier import Make_Amplifier_Typ_II_UpDown
-from .white_cell import Make_White_Cell
-from .compressor import Make_Compressor
